Move AsideWrapper into components/nav folder
